+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## HEAD (unreleased)
 
+Authors: Maxim Polunin
+
+* fix: support nested temporary global listeners
+
 ## 2.0.0 (7th Mar 2017)
 
 Authors: Sergey Mostovoy, Andrew Kozin, Kyle Tolle, Martin, Rob Miller, Mike